requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.8.1
    # via flask
click==8.1.7
    # via
    #   click-default-group
    #   flask
    #   sqlite-utils
click-default-group==1.2.4
    # via sqlite-utils
flask==3.0.3
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
pillow==10.3.0
    # via -r requirements.in
pluggy==1.5.0
    # via sqlite-utils
python-dateutil==2.9.0.post0
    # via sqlite-utils
six==1.16.0
    # via python-dateutil
sqlite-fts4==1.0.3
    # via sqlite-utils
sqlite-utils==3.36
    # via -r requirements.in
tabulate==0.9.0
    # via sqlite-utils
tqdm==4.66.4
    # via -r requirements.in
wcag-contrast-ratio==0.9
    # via -r requirements.in
werkzeug==3.0.3
    # via flask
